requirements.txt
Flask
sklearn
pandas
numpy
matplotlib
PyYAML
-e .